+++ packages/analytics-client-common/src/index.ts
@@ -8,3 +8,4 @@ export { FetchTransport } from './transports/fetch';
 export { getAnalyticsConnector } from './analytics-connector';
 export { IdentityEventSender } from './plugins/identity';
 export { getLanguage } from './language';
+export { BASE_CAMPAIGN } from './attribution/constants';