+++ packages/analytics-client-common/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-client-common@0.1.0...@amplitude/analytics-client-common@0.2.0) (2022-09-22)
+
+
+### Features
+
+* add campaign params to page view events ([#216](https://github.com/amplitude/Amplitude-TypeScript/issues/216)) ([c0f99b9](https://github.com/amplitude/Amplitude-TypeScript/commit/c0f99b98d0d2c24f6f9486312b568194c690a202))
+
+
+
+
+
 # [0.1.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-client-common@0.0.1...@amplitude/analytics-client-common@0.1.0) (2022-09-16)