+++ packages/analytics-types/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.1](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-types@0.10.0...@amplitude/analytics-types@0.10.1) (2022-09-22)
+
+**Note:** Version bump only for package @amplitude/analytics-types
+
+
+
+
+
 # [0.10.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-types@0.9.0...@amplitude/analytics-types@0.10.0) (2022-09-16)